Initial Commit
OK so The program is using the Qt5WebEngine library
along with the libmpdclient library.
The basic interface is nothing to write home about, but is functional.
hiburanBaru currently doesn't store the playlist locally, and just
expects the MPD server to handle that.
I have created my own version of a jinja templating system to work with
html templates, so users will eventually be able to create their own
skins.
I haven't had a chance to sort out networking from the interface yet,
but it's on my todo list.
This is going to be a full screen app, for use with a raspberry pi
in my car.